+++ TnT/Projects/Presentation/Sources/Coordinator/AppFlow/AppFlowCoordinatorView.swift
@@ -17,25 +17,31 @@ public struct AppFlowCoordinatorView: View {
     }
 
     public var body: some View {
-        Group {
-            if let userType = store.userType {
-                switch userType {
-                case .trainee:
-                    if let store = store.scope(state: \.traineeMainState, action: \.subFeature.traineeMainFlow) {
-                        TraineeMainFlowView(store: store)
+        ZStack {
+            Group {
+                if let userType = store.userType {
+                    switch userType {
+                    case .trainee:
+                        if let store = store.scope(state: \.traineeMainState, action: \.subFeature.traineeMainFlow) {
+                            TraineeMainFlowView(store: store)
+                        }
+                    case .trainer:
+                        if let store = store.scope(state: \.trainerMainState, action: \.subFeature.trainerMainFlow) {
+                            TrainerMainFlowView(store: store)
+                        }
                     }
-                case .trainer:
-                    if let store = store.scope(state: \.trainerMainState, action: \.subFeature.trainerMainFlow) {
-                        TrainerMainFlowView(store: store)
+                } else {
+                    if let store = store.scope(state: \.onboardingState, action: \.subFeature.onboardingFlow) {
+                        OnboardingFlowView(store: store)
                     }
-                }
-            } else {
-                if let store = store.scope(state: \.onboardingState, action: \.subFeature.onboardingFlow) {
-                    OnboardingFlowView(store: store)
                 }
             }
+            .animation(.easeInOut, value: store.userType)
+            
+            // Overlay
+            OverlayContainer()
+                .environmentObject(OverlayManager.shared)
         }
-        .animation(.easeInOut, value: store.userType)
         .onAppear {
             store.send(.onAppear)
         }
